extract out kazhttp.py as a library, write a c++ proxy server that can intercept requests
NEXT we'll use the c++ proxy server to capture and replay requests and debug timeouts.  we can also use it to generate requests and fuzz-test the c++ implementation.  we can also use it to compare our server against a reference to find bugs.

the c++ proxy server uses select to swap between the socket to the browser (client socket, created from the server socket) and the socket to the backend, which is encrypted with an SSL connection.

TODO see if we need to add client certs to that somehow.