fixed
when a new worker is created, the current_result is undefined and the worker is not done.

then immediately, search subscribes to the worker.  on subscription, the worker checks if it's done and just returns the result, and otherwise waits for the result to propagate the next time it works on it.

it's not done, so it doesn't call the subscriber with the current result.

actually i have no idea.