actually fix immediate search.  FIX BUG message cacher didn't re-run on cache changes when the cache was still valid because [version numbers match / changes were only made [locally / within the same tab]]
the fix was just always rerunning the message caching worker when we're ensuring a valid cache.  we should probably only do it when we've actually _changed_ something in the cache, but we can assume that if we're ensuring the cache is valid it probably changed.

also, the message list cacher is pretty fast to run when the rewrite cache is available because all it requires is gathering the already-parsed messages.  it seems pretty snappy.